+++ CHANGELOG.md
@@ -116,3 +116,4 @@
 - [2026-04-04 14:30] fix(state): prevent double tap duplicate action in state
 - [2026-04-04 14:48] fix(i18n): handle null assertion failure in i18n parser
 - [2026-04-04 15:13] feat(core): add async core stream provider
+- [2026-04-04 15:19] feat(navigation): add async navigation stream provider